requirements.txt
requests~=2.32.5
pytest == 9.0.2
allure-pytest == 2.15.3
allure-python-commons